Use merge-config.sh for integrating KERNEL_ADDITIONAL_CONFIG.
Using 'cat' to combine defconfig fragments is not, in general,
safe. Overriding options is not correctly handled, and it's
possible to produce illegal configurations which violate Kconfig
dependency constraints.

Instead, use the merge-config.sh script from the kernel tree
which more intelligently combines defconfig fragments. An
'alldefconfig' make target is also introduced, as merge-config.sh
depends on it.

Change-Id: I91a8c5e4bcf7072a3f95651dffe17380174134e4