Allow building DTBO images.
taimen uses the DTBO image as an DTB Overlay in order to use the same defconfig for both, taimen and wahoo.
The dtbo.img is built with "make dtbo.img". Allow automatically building it when setting TARGET_NEEDS_DTBOIMAGE to true, or by manually calling make dtboimage.

Thanks to luca020400 for help with the kernel task logic.

[raymanfx] Add AVB hash footer support.

Change-Id: Iaddafaa98e3e84880462a2c35b899e3fea21e161